+++ CHANGELOG.md
@@ -162,6 +162,7 @@
 * Added flag to pre-activate entity placeholders in world scenes
 * Weapon ViewModel can be forced to always or never wait for lock time to do effect [Experimental]
 * Added subtitle convenience methods from Shattered 2 [Experimental]
+* Added fade canvasgroup convenience methods [Experimental]
 * Extended projectile and explosion handling for weapons [WIP]
 * Extended effects handling for sprite weapon view models [WIP]
 * Melee weapon damage can be delayed [WIP]